% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjusted_mixtures}
\alias{adjusted_mixtures}
\title{Adjusted mixtures}
\format{
An object of class \code{list} of length 30000.
}
\usage{
adjusted_mixtures
}
\description{
Adjusted mixtures
}
\keyword{datasets}
