% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eps_max}
\alias{eps_max}
\title{Upper bound approximation}
\format{
An object of class \code{numeric} of length 30000.
}
\usage{
eps_max
}
\description{
Upper bound approximation
}
\keyword{datasets}
