% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eps_min}
\alias{eps_min}
\title{Lower bound approximation}
\format{
An object of class \code{numeric} of length 30000.
}
\usage{
eps_min
}
\description{
Lower bound approximation
}
\keyword{datasets}
