% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{original_mixtures}
\alias{original_mixtures}
\title{Original mixtures}
\format{
An object of class \code{list} of length 30000.
}
\usage{
original_mixtures
}
\description{
Original mixtures
}
\keyword{datasets}
