% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{posterior_summaries}
\alias{posterior_summaries}
\title{Computation of posterior summaries}
\usage{
posterior_summaries(model_out)
}
\arguments{
\item{model_out}{Object created by \code{sample_PM}, \code{sample_PM} or \code{burn_thin}.}
}
\description{
Providing as input an object created by \code{sample_PM}, \code{sample_PMM} or \code{burn_thin}, the function computes posterior summaries.
}
