% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_PM.R
\name{sample_PM}
\alias{sample_PM}
\title{Draw posterior samples from a Bayesian Poisson regression model}
\usage{
sample_PM(
  y,
  X,
  offset = rep(1, length(y)),
  prior_precision_beta = 0.001,
  algorithm = c("Automatic", "IAMS", "MH-IAMS", "RIAMS"),
  n_iter = 10000,
  print = n_iter/10,
  thin = 1,
  T1 = 500,
  T2 = 250,
  pL = 0.05,
  pU = 0.05
)
}
\arguments{
\item{y}{Vector of observed responses.}

\item{X}{Matrix with fixed effects. The intercept is added by the function.}

\item{offset}{Vector with possible offsets. Default all 1 values.}

\item{prior_precision_beta}{Precision of the Gaussian prior of the regression coefficients.}

\item{algorithm}{Choice of the algorithm: "Automatic", "IAMS", "MH-IAMS" and "RIAMS".}

\item{n_iter}{Number of MCMC iterations.}

\item{print}{Frequency of iterations for printing the progress.}

\item{thin}{Number of thinned observations.}

\item{T1}{Number of iterations with IAMS algorithm for initializing the algorithm (used in "Automatic", "MH-IAMS" and "RIAMS")}

\item{T2}{Number of iterations in which the need for mixture adjustment ("Automatic" and "RIAMS") and MH step ("Automatic") is verified.}

\item{pL}{Threshold for the proportion of problematic approximations to activate the MH step ("Automatic").}

\item{pU}{Threshold for the proportion of problematic approximations to activate the mixture adjustment ("Automatic").}
}
\value{
The function returns a list with:
\itemize{
\item \code{algorithm}.
\item \code{out}: List with posterior samples of regression coefficients.
\item \code{elapsed_time}.
\item \code{accept_beta}: acceptance rate for the vector of regression coefficients (except for "IAMS" algorithm).
\item \code{kU}: for each residual, it contains the proportion of preliminary iterations falling in the upper tail ("RIAMS" and "Automatic" algorithms).
\item \code{kL}: for each residual, it contains the proportion of preliminary iterations falling in the lower tail ("Automatic" algorithm).
}
}
\description{
Allow the user to specify a Bayesian Poisson regression model and draw samples from the posterior distributions of parameters through different algorithms.
}
