% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_PMM.R
\name{sample_PMM}
\alias{sample_PMM}
\title{Draw posterior samples from a Bayesian linear mixed model}
\usage{
sample_PMM(
  y,
  X,
  offset = rep(1, length(y)),
  reff_list,
  a_pri_s2g,
  b_pri_s2g,
  prior_precision_beta = 0.001,
  beta_init = rep(0, ncol(X)),
  s2g_init = rep(1, length(reff_list)),
  algorithm = c("Automatic", "IAMS", "MH-IAMS", "RIAMS"),
  n_iter = 10000,
  pr = n_iter/10,
  thin = 1,
  T1 = 500,
  T2 = 250,
  pL = 0.05,
  pU = 0.05
)
}
\arguments{
\item{y}{Vector of observed responses.}

\item{X}{Matrix with fixed effects. The intercept is added by the function.}

\item{offset}{Vector with possible offsets. Default all 1 values.}

\item{reff_list}{List of lists containing information on random effects. For each term, the list must contain: a design matrix \code{Z} with a number of columns \code{m_j}, a precision matrix \code{K} with its rank \code{rank_K} and a logical \code{K_full_rank} (\code{TRUE} if \code{K} is full rank), and the possible linear constraint defined by a matrix \code{A} and a vector \code{e}.}

\item{a_pri_s2g}{Vector with the shape parameters of the Inverse-Gamma priors for the random effects variances.}

\item{b_pri_s2g}{Vector with the scale parameters of the Inverse-Gamma priors for the random effects variances.}

\item{prior_precision_beta}{Precision of the Gaussian prior of the regression coefficients.}

\item{beta_init}{Initial values for the vector of regression coefficients.}

\item{s2g_init}{Initial values for the random effects variances.}

\item{algorithm}{Choice of the algorithm: "Automatic", "IAMS", "MH-IAMS" and "RIAMS".}

\item{n_iter}{Number of MCMC iterations.}

\item{pr}{Frequency of iterations for printing the progress.}

\item{thin}{Number of thinned observations.}

\item{T1}{Number of iterations with IAMS algorithm for initializing the algorithm (used in "Automatic", "MH-IAMS" and "RIAMS")}

\item{T2}{Number of iterations in which the need for mixture adjustment ("Automatic" and "RIAMS") and MH step ("Automatic") is verified.}

\item{pL}{Threshold for the proportion of problematic approximations to activate the MH step ("Automatic").}

\item{pU}{Threshold for the proportion of problematic approximations to activate the mixture adjustment ("Automatic").}
}
\value{
The function returns a list with:
\itemize{
\item \code{algorithm}.
\item \code{out}: List with posterior samples of regression coefficients.
\item \code{elapsed_time}.
\item \code{accept_beta}: acceptance rate for the vector of regression coefficients (except for "IAMS" algorithm).
\item \code{accept_gamma}: acceptance rates for the vectors of random effects (except for "IAMS" algorithm).
\item \code{kU}: for each residual, it contains the proportion of preliminary iterations falling in the upper tail ("RIAMS" and "Automatic" algorithms).
\item \code{kL}: for each residual, it contains the proportion of preliminary iterations falling in the lower tail ("Automatic" algorithm).
}
}
\description{
Allow the user to specify a Bayesian linear mixed model, with flexibility in the prior choices.
}
